backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# File Upload Support
python-multipart==0.0.6

# Image Processing and Face Recognition
opencv-python==4.10.0.84
face-recognition==1.3.0
dlib==19.24.6
Pillow==10.4.0

# Database and Storage
supabase==2.0.3

# Environment Variables
python-dotenv==1.0.0

# Additional Dependencies
numpy==2.0.2

# Authentication
bcrypt==4.1.1
PyJWT==2.8.0
email-validator==2.2.0
